Converted TypedBasedNotificationTests
